Equilibrium initialisation list creation

equilibrium_init_create creates an equilibrium initialisation state to be used within later model runs

equilibrium_init_create(
  age_vector,
  het_brackets,
  country = NULL,
  admin_unit = NULL,
  ft,
  EIR,
  model_param_list
)

Arguments

age_vector
Vector of age brackets.

het_brackets
Integer number of biting heteogenity compartments.

country
String for country of interest. If NULL the seasonal parameters will attempt to be loaded using just the admin unit, however if there is ambiguity in the admin unit an error will be thrown. If both NULL then no seasonality is assumed. Default = NULL.

admin_unit
String for admin unit with country for loading seasonal parameters. If country is NULL, the admin unit will attempt to be located,however if there is ambiguity in the admin unit an error will be thrown. If both country and admin_unit are NULL then no seasonality is assumed. Default = NULL.

ft
Numeric for the frequency of people seeking treatment.

EIR
Numeric for desired annual EIR.

model_param_list
List of epidemiological parameters created by