Model Parameter List Creation

model_param_list_create creates list of model parameters to be used within equilibrium_init_create

model_param_list_create(
  eta = 1/(21 * 365),
  rho = 0.85,
  a0 = 2920,
  sigma2 = 1.67,
  max_age = 100 * 365,
  rA = 1/195,
  rT = 0.2,
  rD = 0.2,
  rU = 1/110.299,
  rP = 1/15,
  dE = 12,
  delayGam = 12.5,
  cD = 0.0676909,
  cT = 0.322 * cD,
  cU = 0.006203,
  gamma1 = 1.82425,
  d1 = 0.160527,
  dID = 3650,
  ID0 = 1.577533,
  kD = 0.476614,
  uD = 9.44512,
  aD = 8001.99,
  fD0 = 0.007055,
  gammaD = 4.8183,
  alphaA = 0.75735,
  alphaU = 0.185624,
  b0 = 0.590076,
  b1 = 0.5,
  dB = 3650,
  IB0 = 43.8787,
  kB = 2.15506,
  uB = 7.19919,
  phi0 = 0.791666,
  phi1 = 0.000737,
  dCA = 10950,
  IC0 = 18.02366,
  kC = 2.36949,
  uCA = 6.06349,
  PM = 0.774368,
  dCM = 67.6952,
  delayMos = 10,
  tau1 = 0.69,
  tau2 = 2.31,
  mu0 = 0.132,
  Q0 = 0.92,
  chi = 0.86,
  bites_Bed = 0.89,
  bites_Indoors = 0.97,
  muEL = 0.0338,
  muLL = 0.0348,
  muPL = 0.249,
  dEL = 6.64,
  dLL = 3.72,
  dPL = 0.643,
  gammaL = 13.25,
  km = 11,
  cm = 0.05,
  betaL = 21.2,
  num_int = 1,
  itn_cov = 0,
  irs_cov = 0,
  ITN_IRS_on = -1,
  DY = 365,
  d_ITN0 = 0.41,
  r_ITN0 = 0.56,
  r_ITN1 = 0.24,
  r_IRS0 = 0.6,
  d_IRS0 = 1,
  irs_half_life = 0.5 * DY,
  itn_half_life = 2.64 * DY,
  IRS_interval = 1 * DY,
  ITN_interval = 3 * DY,
  ...
)

Arguments

eta

Death rate for expoential population distribtuion, i.e. 1/Mean Population Age. Default = 0.0001305

rho

Age-dependent biting parameter. Default = 0.85

a0

Age-dependent biting parameter. Default = 2920

sigma2

Variance of the log heterogeneity in biting rates. Default = 1.67

max_age

Maximum age in days. Default = 100*365

rA

Rate of leaving asymptomatic infection. Default = 0.00512821

rT

Rate of leaving treatment. Default = 0.2

rD

Rate of leaving clinical disease. Default = 0.2

rU

Rate of recovering from subpatent infection. Default = 0.00906627

rP

Rate of leaving prophylaxis. Default = 0.06666667

dE

Latent period of human infection. Default = 12

delayGam

Lag from parasites to infectious gametocytes. Default = 12.5

cD

Untreated disease contribution to infectiousness. Default = 0.0676909

cT

Treated disease contribution to infectiousness. Default = 0.322 * cD

cU

Subpatent disease contribution to infectiousness. Default = 0.006203

gamma1

Parameter for infectiousness of state A. Default = 1.82425

d1

Minimum probability due to maximum immunity. Default = 0.160527

dID

Inverse of decay rate. Default = 3650

ID0

Scale parameter. Default = 1.577533

kD

Shape parameter. Default = 0.476614

uD

Duration in which immunity is not boosted. Default = 9.44512

aD

Scale parameter relating age to immunity. Default = 8001.99

fD0

Time-scale at which immunity changes with age. Default = 0.007055

gammaD

Shape parameter relating age to immunity. Default = 4.8183

alphaA

PCR detection probability parameters state A. Default = 0.757

alphaU

PCR detection probability parameters state U. Default = 0.186

b0

Maximum probability due to no immunity. Default = 0.590076

b1

Maximum relative reduction due to immunity. Default = 0.5

dB

Inverse of decay rate. Default = 3650

IB0

Scale parameter. Default = 43.8787

kB

Shape parameter. Default = 2.15506

uB

Duration in which immunity is not boosted. Default = 7.19919

phi0

Maximum probability due to no immunity. Default = 0.791666

phi1

Maximum relative reduction due to immunity. Default = 0.000737

dCA

Inverse of decay rate. Default = 10950

IC0

Scale parameter. Default = 18.02366

kC

Shape parameter. Default = 2.36949

uCA

Duration in which immunity is not boosted. Default = 6.06349

PM

New-born immunity relative to mother’s. Default = 0.774368

dCM

Inverse of decay rate of maternal immunity. Default = 67.6952

delayMos

Extrinsic incubation period. Default = 10

tau1

Duration of host seeking, assumed to be constant between species. Default = 0.69

tau2

Duration of mosquito resting after feed. Default = 2.31

mu0

Daily mortality of adult mosquitos. Default = 0.132

Q0

Anthrophagy probability. Default = 0.92

chi

Endophily probability. Default = 0.86

bites_Bed

Percentage of bites indoors and in bed. Default = 0.89

bites_Indoors

Percentage of bites indoors . Default = 0.97

muEL

Per capita daily mortality rate of early stage larvae (low density). Default = 0.0338

muLL

Per capita daily mortality rate of late stage larvae (low density). Default = 0.0348

muPL

Per capita daily mortality rate of pupae. Default = 0.249

dEL

Development time of early stage larvae. Default = 6.64

dLL

Development time of late stage larvae. Default = 3.72

dPL

Development time of pupae. Default = 0.643

gammaL

Relative effect of density dependence on late instars relative to early instars. Default = 13.25

km

Seasonal carrying capacity. Default = 11

cm

Seasonal birth rate. Default = 0.05

betaL

Number of eggs laid per day per mosquito. Default = 21.2

num_int

Number of intervention parameters. Default = 4

itn_cov

The proportion of people that use an ITN. Default = 0

irs_cov

The proportion of people living in houses that have been sprayed. Default = 0

ITN_IRS_on

Time of ITN and IRS to be activated. Default = -1, i.e. never.

DY

Duration of year (days). Default = 365

d_ITN0

Probability of dying with an encounter with ITN (max). Default = 0.41

r_ITN0

Probability of repeating behaviour with ITN (max). Default = 0.56

r_ITN1

Probability of repeating behaviour with ITN (min). Default = 0.24

r_IRS0

Probability of repeating behaviour with IRS (min). Default = 0.6

d_IRS0

Probability of dying with an encounter with IRS (max). Default = 1

irs_half_life

IRS half life. Default = 0.5 * DY

itn_half_life

ITN half life. Default = 2.64 * DY

IRS_interval

How long before IRS is repeated, i.e. when IRS decay = 1. Default = 1 * DY

ITN_interval

How long before ITN is repeated, i.e. when IRS decay = 1. Default = 3 * DY

...

Any other parameters needed for non-standard model. If they share the same name as any of the defined parameters model_param_list_create will stop. You can either write any extra parameters you like individually, e.g. model_param_list_create(extra1 = 1, extra2 = 2) and these parameteres will appear appended to the returned list, or you can pass explicitly the ellipsis argument as a list created before, e.g. model_param_list_create(...=list(extra1 = 1, extra2 = 2))